refactor: All internal extensions moved to package `internal`
