+++ infrastructure/src/main/kotlin/KotlinLibraryPlugin.kt
@@ -1,6 +1,7 @@
 package com.redmadrobot.build
 
 import com.redmadrobot.build.extension.TestOptions
+import com.redmadrobot.build.internal.*
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.tasks.testing.Test
